chore(audit01): fix pre-existing baseline lint failures
Phase 00 preflight requires `cargo fmt`, `clippy -D warnings`, `cargo
test`, and `shellcheck` all green before any further phase lands. The
following pre-existing failures were found at HEAD of `dux-amq-setup`
(none flagged by audit01.md but blocking the preflight gate):

- `src/app/render.rs:1354,1359`: clippy 1.92's `unnecessary_cast` lint
  fires on `as usize` against fields that are already `usize`
  (`scrollback_total`, `scrollback_offset` in `pty.rs`). Drop the casts.
- `dux-amq/install.sh:102-104`: shellcheck SC2015 — rewrite
  `cmd && ok || warn` as an explicit if/else so `warn` only runs when
  the pipeline truly fails, not when `ok` itself returns non-zero.
- `dux-amq/scripts/finalize-claude-migration.sh:10`: shellcheck SC2155 —
  split `local bak=...$(date)` into declare-then-assign so a future
  failure of `date` doesn't get masked by `local`'s exit status.

After these fixes:
  cargo fmt --check                                  OK
  cargo clippy --all-targets -- -D warnings          OK
  cargo test                                         712 passed, 0 failed
  shellcheck install.sh wrappers/* scripts/*.sh      OK

+++ src/app/render.rs
@@ -1351,12 +1351,12 @@ impl App {
                 // navigation feedback here.
                 {
                     use ratatui::widgets::{Scrollbar, ScrollbarOrientation, ScrollbarState};
-                    let total = self.snapshot_buf.scrollback_total as usize;
+                    let total = self.snapshot_buf.scrollback_total;
                     let visible = term_area.height as usize;
                     if total > 0 && visible > 0 && term_area.width >= 2 {
                         // ratatui position is 0 at the top, content_length-1 at the bottom.
                         // dux's offset is 0 when at the latest (bottom): position = total - offset.
-                        let offset = (self.snapshot_buf.scrollback_offset as usize).min(total);
+                        let offset = self.snapshot_buf.scrollback_offset.min(total);
                         let position = total.saturating_sub(offset);
                         let mut state = ScrollbarState::new(total + visible)
                             .viewport_content_length(visible)
